MSN Chinese net
This is a MSN Chinese net that  I do today. It is a beautiful site and
there are many special effects in it. It has two slides which is
difficult , so I spend much time doing it. I should remember them and
use them in the future.
Key code:
It does not finish ,I will do it tomorrow.
//普通的幻灯片实现代码
var index=1;
function slide(){
var leftL=(index-1)*824;
$("#main1CLeftUl").animate({"left":"-"+leftL+"px"},500);
$(".main1NumA").css({"top":"0", "background-color":"#fff", "color":
"#000"});
$(".main1NumA"+index).css({"top":"-10px", "background-color":"#000",
"color": "#fff"});
index++;
if (index==5) {
index=1;
};
}
//漂亮的循环幻灯片
var indexBeauty=1;
function slideBeauty(){
for( i=1; i<=8;i++)
{
if (i>indexBeauty) {
$("#main3PicliID"+i).attr("class","main3Picli"+(i-indexBeauty));
}else{
$("#main3PicliID"+i).attr("class","main3Picli"+(8+i-indexBeauty));
}
}
indexBeauty++;
if (indexBeauty==9) {
indexBeauty=1;
};
}